Changes after the review of Mr.Raimondi: -Added a Utility class. -Added better specs for interface. -Added some comments for decrypt and encrypt methods
