+++ app/src/main/java/ingsw/group1/findmyphone/cryptography/SMSCipher.java
@@ -6,6 +6,9 @@ import androidx.annotation.NonNull;
 import com.eis.smslibrary.SMSMessage;
 import com.eis.smslibrary.SMSPeer;
 
+import static ingsw.group1.findmyphone.cryptography.SMSCipherUtils.addPadding;
+import static ingsw.group1.findmyphone.cryptography.SMSCipherUtils.fromStringToAscii;
+
 /**
  * Class used to cipher and decipher a SMSMessage.
  *
@@ -13,7 +16,6 @@ import com.eis.smslibrary.SMSPeer;
  */
 public class SMSCipher implements MessageCipher<SMSMessage> {
 
-    private static final char PARSING_CHARACTER = '0';
     private String passwordToCypher;
 
     /**
@@ -60,12 +62,15 @@ public class SMSCipher implements MessageCipher<SMSMessage> {
      * @return The encrypted String.
      */
     public static String encrypt(String stringToCipher, String key) {
-        int[] strInt = SMSCipher.fromStringToAscii(stringToCipher);
-        int[] keyInt = SMSCipher.fromStringToAscii(key);
+        //converts every characters of both string into ASCII Code and saved inside a int array
+        int[] strInt = fromStringToAscii(stringToCipher);
+        int[] keyInt = fromStringToAscii(key);
         StringBuilder encryptedStr = new StringBuilder();
         for (int i = 0; i < strInt.length; i++) {
+            //XOR the pair of characters of index i of the 2 strings and convert the result in hex
             String s = Integer.toHexString(strInt[i] ^ keyInt[i % keyInt.length]);
-            String s2 = SMSCipher.addPadding(s, 2);
+            //this padding is add because for instance 07 is written 7 inside the string
+            String s2 = addPadding(s, 2);
             encryptedStr.append(s2);
         }
         return encryptedStr.toString();
@@ -80,15 +85,19 @@ public class SMSCipher implements MessageCipher<SMSMessage> {
      */
     public static String decrypt(String stringToDecipher, String key) {
         StringBuilder decrypted = new StringBuilder();
-        int[] keyInt = SMSCipher.fromStringToAscii(key);
+        //converts every characters of the key into ASCII Code and saves inside a int array
+        int[] keyInt = fromStringToAscii(key);
         int[] intArr = new int[stringToDecipher.length()];
         String[] strArr = new String[stringToDecipher.length() / 2];
         for (int i = 0; i < strArr.length; i++) {
+            //save in strArr[i] every couple of 2 characters because in hex a couple of value is a character 01 23 45 ...
             strArr[i] = stringToDecipher.substring(i * 2, (i * 2) + 2);
         }
         for (int i = 0; i < strArr.length; i++) {
+            //converts the strArr (which was in hex) into decimal.
             int decimal = Integer.parseInt(strArr[i], 16);
             intArr[i] = decimal;
+            //Decrypt the String and converts it from ascii to a normal String.
             String s = String.valueOf(intArr[i] ^ keyInt[i % keyInt.length]);
             int n = Integer.valueOf(s, 10);
             decrypted.append((char) n);
@@ -96,38 +105,6 @@ public class SMSCipher implements MessageCipher<SMSMessage> {
         return decrypted.toString();
     }
 
-    /**
-     * Converts an integer to a string and adds to it the padding character.
-     * If stringToPad.length is grater or equals to @length than it only converts the intToPadd
-     * into a string.
-     *
-     * @param stringToPad The String that need the padding.
-     * @param length      The wanted length for the string.
-     * @return String with the correct length.
-     */
-    public static String addPadding(String stringToPad, int length) {
-        String paddedString = stringToPad;
-        while (paddedString.length() < length) {
-            paddedString = PARSING_CHARACTER + paddedString;
-        }
-        return paddedString;
-    }
-
-    /**
-     * Converts every characters of the string into integer of Ascii table and save their values
-     * in an array of integer.
-     *
-     * @param stringToBeConverted The string to be converted.
-     * @return An array of integer containing in position i the value of the character i of the
-     * string.
-     */
-    public static int[] fromStringToAscii(String stringToBeConverted) {
-        int[] integerArray = new int[stringToBeConverted.length()];
-        for (int i = 0; i < stringToBeConverted.length(); i++) {
-            integerArray[i] = (int) stringToBeConverted.charAt(i);
-        }
-        return integerArray;
-    }
 
     /**
      * Sets the password of the cypher object.
